Making progress, trying to start SearchableActivity causes inflate exception, error inflating class fragment
