+++ app/build.gradle
@@ -63,4 +63,7 @@ dependencies {
     testImplementation "androidx.room:room-testing:$room_version"
 
     implementation 'androidx.cardview:cardview:1.0.0'
+
+    // viewmodel and livedata /app/build.gradle dependencies
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.1.0'
 }
