+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.CST438_Project01_Starter">
         <activity
+            android:name=".SearchableActivity"
+            android:exported="false" />
+        <activity
             android:name=".HomeActivity"
             android:exported="false" />
         <activity
