+++ app/src/main/java/com/daclink/drew/sp22/cst438_project01_starter/MealSearchResultsAdapter.java
@@ -21,7 +21,7 @@ public class MealSearchResultsAdapter extends RecyclerView.Adapter<MealSearchRes
     @NonNull
     @Override
     public MealSearchResultsHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.activity_search, parent, false);
+        View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.activity_searchbak, parent, false);
         return new MealSearchResultsHolder(itemView);
     }
