+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,26 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
+    app:startDestination="@id/searchableFragment">
 
     <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.daclink.drew.sp22.cst438_project01_starter.FirstFragment"
+        android:id="@+id/searchableFragment"
+        android:name="com.daclink.drew.sp22.cst438_project01_starter.SearchableFragment"
         android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
-
-        <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.daclink.drew.sp22.cst438_project01_starter.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
-
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
-    </fragment>
+        tools:layout="@layout/fragment_searchable"/>
 </navigation>
